+++ components/logical-operations/LogicalOperations.js
@@ -3,12 +3,10 @@
 import BitDisplay         from '../BitDisplay';
 import BitButtonGroup     from '../logical-operations/BitButtonGroup';
 import BitOperationsGroup from '../logical-operations/BitOperationsGroup';
+import SingleBitGroup     from '../logical-operations/SingleBitGroup';
 
 /** States of single bits */
-let firstBitgroup   = [1, 0, 1, 0, 1, 0, 1, 0];
-let secondBitgroup  = [0, 1, 0, 1, 0, 1, 0, 1];
 let result          = [0, 0, 0, 0, 0, 0, 0, 0];
-
 
 let CreateCaption = props => (<h3>{props.topic}</h3>);
